+++ menu-emails/public/cancelacionCreditoSASE/index.html
@@ -1,0 +1,180 @@
+<!DOCTYPE html
+  PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
+<html xmlns="http://www.w3.org/1999/xhtml">
+
+<head>
+  <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <title>Entrega en Tienda Pago con Tarjeta</title>
+
+</head>
+
+<style type="text/css">
+  body {
+    margin: 0;
+    padding: 0;
+    font-size: 16px;
+    line-height: 20px;
+    font-family: "HelveticaNeue-Light", "Helvetica Neue Light", "Helvetica Neue", Helvetica, Arial, "Lucida Grande",
+      sans-serif;
+    color: #333333;
+  }
+
+  a {
+    text-decoration: none;
+    cursor: pointer;
+  }
+
+  p {
+    margin: 0px;
+    padding: 0px;
+    color: #333333;
+  }
+</style>
+
+<body>
+  <table align="center" border="0" cellpadding="0" cellspacing="0" style="border-collapse: collapse; max-width: 600px; min-width: 375px; border: 1px solid #eaeaea;
+    table-layout: auto; color: #333333" bgcolor="#ffffff">
+    <tr>
+      <td>
+        <img width="100%" src="https://ekt-atap.s3.amazonaws.com/ekt-mailings-transaccionales/600.png" alt="" />
+      </td>
+    </tr>
+    <tr>
+      <td
+        style="padding-top:23px; background-image:'https://ekt-atap.s3.amazonaws.com/ekt-mailings-transaccionales/semi-circle.png'; background-repeat:repeat;">
+        <table id="names" border="0" cellpadding="0" cellspacing="0" style="border-collapse: collapse;" max-width="458">
+          <tr>
+            <td>
+              <img src="https://ekt-atap.s3.amazonaws.com/ekt-mailings-transaccionales/logo-elektra_izq.png" width="185"
+                alt="" />
+            </td>
+          </tr>
+          <tr>
+            <td style="font-size: 18px; font-weight: 600; padding-left: 20px; padding-top: 26px; padding-bottom: 15px;">
+              Hola Rene,</td>
+          </tr>
+          <tr>
+            <td valign="middle" style="padding-left: 20px; width: 320px">
+              <p style="font-size: 18px; line-height: 24px;  margin:0px;">
+                <span>Tu pedido</span>
+                <span style="color:#fe1d00; font-weight: 600; text-decoration:none;"><a
+                    style="color:#fe1d00; font-weight: 600; text-decoration:none;">v12121212ekt-01</a> </span>
+              </p>
+              <p style="font-size: 18px; line-height: 24px; margin: 0px">
+                <span>ha sido</span>
+                <span style="color:#fe1d00; font-weight: 600;">cancelado</span>
+              </p>
+            </td>
+          </tr>
+        </table>
+      </td>
+    </tr>
+    <td style="padding:0 20px">
+      <table cellspacing="0" style="border-collapse: collapse; max-width: 560px; color: #333333;">
+        <tr>
+          <td style="font-size: 16px;">
+            <p style="margin: 17px 0 12px 0px;">
+              <span>También hemos <span style="color:#00af2e;">restablecido tu capacidad
+                  depago con éxito</span></span>
+            </p>
+          </td>
+        </tr>
+      </table>
+    </td>
+    </tr>
+    <tr>
+    <tr>
+      <td style="padding:0 20px">
+        <p style="margin: 40px 0 15px 0px;">Te invitamos a que ingreses nuevamente al sitio y conozcas nuestra gran
+          variedad de productos.</p>
+        <p style="margin: 20px 0 40px 0px;">¡Te esperamos en <span><a style="color:#fe1d00;"
+              href="https://www.elektra.com.mx/">www.elektra.com.mx</a></span>!
+        </p>
+      </td>
+    </tr>
+    <tr>
+
+    </tr>
+    <tr>
+      <td align="center" bgcolor="#fedd00">
+        <table border="0" cellpadding="0" cellspacing="0" style="font-size: 14px; padding-left:30px" width="100%">
+          <tr>
+            <td height="30px" style="font-size: 14px; padding-top: 9px">
+              <span style="font-weight: 600;">Síguenos en</span></td>
+          </tr>
+          <tr style="height: 40px">
+            <th align="left">
+              <img style="padding-right: 17px; width: 25px;height: auto;"
+                src="https://elektra.vteximg.com.br/arquivos/fb_transactional_email.png" alt="">
+              <img style="padding-right: 17px; width: 25px;height: auto;"
+                src="https://elektra.vteximg.com.br/arquivos/twitter_transactional_email.png" alt="">
+              <img style="width: 25px;height: auto;"
+                src="https://elektra.vteximg.com.br/arquivos/instagram_transactional_email.png" alt="">
+
+            </th>
+          </tr>
+          <tr>
+            <td style="font-size: 14px; padding-top: 15px">
+              <span style="font-weight: 600;">Contacto</span></td>
+          </tr>
+          <tr>
+            <td style="font-size: 14px; padding-top: 9px; padding-bottom: 20px">
+              <span style="font-weight: 600;"> <a style="color:#333333" href="tel:5575775547">Tel: 55 7577
+                  5547</a></span>
+              <p>Lunes a domingo de 9:00 am a 9:00 pm</p>
+            </td>
+          </tr>
+        </table>
+      </td>
+    </tr>
+    <tr>
+      <td align="center" bgcolor="#4e4e4e" height="20" style="height: 39px;">
+        <table border="0" cellpadding="0" cellspacing="0" style="font-size: 12px;" width="90%">
+          <tbody>
+            <tr style="height: 39px">
+              <th align="left" width="33%">
+                <p style="margin: 0px; color:white; font-size: 12px; line-height:16px; font-weight: normal;">
+                  <a target="blank" href="https://www.elektra.com.mx/"
+                    style="text-decoration: underline; color: white">Preguntas Frecuentes</a>
+                </p>
+              </th>
+              <td align="center" width="33%">
+                <p style="margin: 0px;  color:white; font-size: 12px; line-height:16px; font-weight: normal;">
+                  <a target="blank" href="https://www.elektra.com.mx/"
+                    style="text-decoration: underline; color: white">Términos de Promociones</a>
+                </p>
+              </td>
+              <th align="center" width="33%">
+                <p style="margin: 0px;  color:white; font-size: 12px; line-height:16px; font-weight: normal;">
+                  <a target="blank" href="https://www.elektra.com.mx/"
+                    style="text-decoration: underline; color: white">Envíos y devoluciones</a>
+                </p>
+              </th>
+            </tr>
+            <tr>
+            </tr>
+          </tbody>
+        </table>
+      </td>
+    </tr>
+    <tr>
+      <td align="center" bgcolor="#fedd00" style="height: 54px;">
+        <table border="0" cellpadding="0" cellspacing="0" style="font-size: 14px;" width="81%">
+          <tbody>
+            <tr>
+              <td style="text-align: center; color:white; font-size: 14px">
+                <p style="margin: 0px; font-size: 14px; font-weight: normal; color:black; text-transform: uppercase;">
+                  ELEKTRA COM, S.A. DE C.V</p>
+                <p style="margin: 0px; font-size: 14px; font-weight: normal; color:black; text-transform: uppercase;">
+                  RFC: OSC990607QHA</p>
+              </td>
+            </tr>
+          </tbody>
+        </table>
+      </td>
+    </tr>
+  </table>
+</body>
+
+</html>
