+++ menu-emails/src/views/misc/path.js
@@ -1,0 +1,140 @@
+const path = [
+  {
+    src: "/pedidoCreadoBazTienda/index.html",
+    url: "/correocreadobaz"
+  },
+  {
+    src: "/pedidoListoEntregaTienda/index.html",
+    url: "/entregaTienda"
+  },
+  {
+    src: "/pedidoCreadoTienda/index.html",
+    url: "/correocreadotiendaTC"
+  },
+  {
+    src: "/pedidoEntregadoTienda/index.html",
+    url: "/correoentregadotienda"
+  },
+  {
+    src: "/pedidoEntregadoCreditoSE/index.html",
+    url: "/correoentregadoCreditoSE"
+  },
+  {
+    src: "/pedidoEntregadoCreditoCE/index.html",
+    url: "/correoentregadoCreditoCE"
+  },
+  {
+    src: "pedidoConfirmadoTiendaOriginacionCredito/index.html",
+    url: "/confirmadotiendacredito"
+  },
+  {
+    src: "pedidoConfirmadoTiendaMercadoPago/index.html",
+    url: "/pedidoConfirmadotiendamercadopago"
+  },
+  {
+    src: "pedidoConfirmadoTiendaPaypal/index.html",
+    url: "/pedidoConfirmadotiendapaypal"
+  },
+  {
+    src: "pedidoConfirmadoTiendaPayU/index.html",
+    url: "/pedidoConfirmadotiendapayu"
+  },
+  {
+    src: "pedidoListoEntregaTiendaEktEnganche/index.html",
+    url: "/entregaTiendaCreditoEktSE"
+  },
+  {
+    src: "pedidoListoEntregaTiendaCE/index.html",
+    url: "/entregaTiendaCreditoEktCE"
+  },
+  {
+    src: "confirmadoDomOneClic/index.html",
+    url: "/confirmadOneClic"
+  },
+  {
+    src: "enviadoOneClic/index.html",
+    url: "/enviadOneClic"
+  },
+  {
+    src: "pedidoEntregadoOneClic/index.html",
+    url: "/pedidoEntregadoOneClic"
+  },
+  {
+    src: "engancheCreditoEktTienda/index.html",
+    url: "/aprobacioncreditotienda"
+  },
+  {
+    src: "pedidoConfirmadoDomPaypal/index.html",
+    url: "/pedidoConfirmadoPayPal"
+  },
+  {
+    src: "pedidoEntregadoDomPayPal/index.html",
+    url: "/pedidoEntregadoPayPal"
+  },
+  {
+    src: "enviadoDomPayPal/index.html",
+    url: "/pedidoEnviadoPayPal"
+  },
+  {
+    src: "pedidoCreadoBazDomRef/index.html",
+    url: "/pedidoCreadoBazRef"
+  },
+  {
+    src: "pedidoEnviadoBazDomRef/index.html",
+    url: "/pedidoEnviadoBazRef"
+  },
+  {
+    src: "pedidoEntregadoBazDomRef/index.html",
+    url: "/pedidoEntregadoBazDomRef"
+  },
+  {
+    src: "pedidoConfirmadoDigitalTDDTDC/index.html",
+    url: "/confirmadoDigitalTDDTDC"
+  },
+  {
+    src: "pedidoListoDigital/index.html",
+    url: "/listoDigitalTDDTDC"
+  },
+  {
+    src: "pedidoCreadoDomItalikaTDDTDC/index.html",
+    url: "/pedidoCreadoItalika"
+  },
+  {
+    src: "pedidoEnviadoDomitalikaTDDTDC/index.html",
+    url: "/pedidoEnviadoItalika"
+  },
+  {
+    src: "pedidoEntregadoDomitalikaTDDTDC/index.html",
+    url: "/pedidoEntregadoItalika"
+  },
+  {
+    src: "pedidoConfirmadoTiedaCE/index.html",
+    url: "/confirmadotiendacreditoCE"
+  },
+  {
+    src: "/creditoCanceladoantesdentregaCE/index.html",
+    url: "/canceladoEngancheAntesdeEntrega"
+  },
+  {
+    src: "/creditoCanceladoSE/index.html",
+    url: "/canceladoEngancheAntesdeEntregaSE"
+  },
+  {
+    src: "/creditoCanceladoCE/index.html",
+    url: "/cancelacionCreditoCE"
+  },
+  {
+    src: "/creditoCanceladoCA/index.html",
+    url: "/pedidoCanceladoAbono"
+  },
+  {
+    src: "/creditoCanceladoCACE/index.html",
+    url: "/pedidoCanceladoAbonoEnganche"
+  },
+  {
+    src: "/cancelacionCreditoSASE/index.html",
+    url: "/cancelacionCreditoSASE"
+  }
+];
+
+export default path;
